requirements.txt
fastapi
uvicorn
python-multipart
jinja2
ffmpeg
pydub
librosa
numpy
pandas
soundfile
tensorflow
wave
python-docx 
docx
Keras-Preprocessing